Resolved all warnings
