+++ .swiftlint.yml
@@ -19,3 +19,6 @@ line_length:
   ignores_function_declarations: true
   ignores_urls: true
   warning: 140
+
+file_length:
+  warning: 500
