+++ Adyen/Utilities/External/AllCountriesPhoneExtensions.swift
@@ -1,12 +1,11 @@
 //
-// Copyright (c) 2022 Adyen N.V.
+// Copyright (c) 2023 Adyen N.V.
 //
 // This file is open source and available under the MIT license. See the LICENSE file for more info.
 //
 
 import Foundation
 
-// swiftlint:disable file_length
 internal let allCountriesPhoneExtensions: [String: String] = [
 
     "AF": "+93",
@@ -492,4 +491,3 @@ internal let allCountriesPhoneExtensions: [String: String] = [
     "VI": "+1 340"
 
 ]
-// swiftlint:enable file_length
